% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{hello}
\alias{hello}
\title{Print hello world}
\usage{
hello(recipient = "world")
}
\arguments{
\item{recipient}{Character. Either "world" or another place.}
}
\value{
This function returns a "hello world" print.
}
\description{
Print hello world
}
\keyword{hello}
\keyword{world}
